Update dotbot installation script and configuration
This commit updates the dotbot installation script and configuration to include the necessary environment variable for running the dotbot-brew plugin. The PATH is now set to include the Homebrew binary directory. Additionally, the Miniconda installation command has been modified to remove the downloaded installation script after installation. The requirements-pipx.txt file has been removed as it is no longer needed. Finally, a new user has been added to the Dockerfile.